+++ myFile.csv
@@ -1,4 +1,4 @@
 id,type,name,status,description,start,end,duration,epic
-1,TASK,Покормить кота,NEW,Кормить влажным кормом,null,null,0
-2,TASK,Помыть посуду,NEW,Помыть тарелки и ложки,null,null,0
-4,SUBTASK,Методы,NEW,Выучить методы,null,null,0,3
+1,TASK,Сдать отчет,IN_PROGRESS,Отчет по работе,2024.05.05,2024.06.05,0
+2,TASK,Выдать премию,NEW,Премия IT-отделу,2024.05.05,2024.06.05,0
+3,EPIC,Изучить доп. материалы,NEW,Изучить лекции,2024.01.05,2024.05.05,0
