+++ src/ru/kedroff/project/exceptionPackage/ManagerSaveException.java
@@ -1,0 +1,8 @@
+package ru.kedroff.project.exceptionPackage;
+
+public class ManagerSaveException extends RuntimeException {
+    public ManagerSaveException(final String message) {
+        super(message);
+    }
+}
+
