+++ src/ru/kedroff/project/exceptionPackage/TaskInputDateException.java
@@ -1,0 +1,7 @@
+package ru.kedroff.project.exceptionPackage;
+
+public class TaskInputDateException extends Exception {
+    public TaskInputDateException(final String message) {
+        super(message);
+    }
+}
