+++ src/ru/kedroff/project/managers/historyManager/HistoryManager.java
@@ -1,0 +1,13 @@
+package ru.kedroff.project.managers.historyManager;
+
+import ru.kedroff.project.tasks.Task;
+
+import java.util.ArrayList;
+
+public interface HistoryManager {
+    void add(Task task);
+
+    void remove(int id);
+
+    ArrayList<Task> getHistory();
+}
