+++ src/ru/kedroff/project/tasks/TaskStatuses.java
@@ -1,0 +1,7 @@
+package ru.kedroff.project.tasks;
+
+public enum TaskStatuses {
+    DONE,
+    IN_PROGRESS,
+    NEW
+}
