+++ src/ru/kedroff/project/tasks/TypeOfTask.java
@@ -1,0 +1,7 @@
+package ru.kedroff.project.tasks;
+
+public enum TypeOfTask {
+    TASK,
+    EPIC,
+    SUBTASK
+}
