+++ src/ru/kedroff/project/utils/Utils.java
@@ -1,0 +1,49 @@
+package ru.kedroff.project.utils;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Utils {
+    public static boolean isNumber(String str) throws NumberFormatException {
+        try {
+            Integer.parseInt(str);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+
+    public static LocalDateTime formattedTime(LocalDateTime localDateTime) {
+        if (localDateTime == null) {
+            return null;
+        }
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm");
+        return LocalDateTime.parse(localDateTime.format(formatter), formatter);
+    }
+
+    public static class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+
+        private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+
+        @Override
+        public void write(JsonWriter jsonWriter, LocalDateTime localDateTime) throws IOException {
+            if (localDateTime == null) {
+                jsonWriter.nullValue();
+            } else {
+                jsonWriter.value(localDateTime.format(DATE_TIME_FORMATTER));
+            }
+        }
+
+        @Override
+        public LocalDateTime read(JsonReader jsonReader) throws IOException {
+            String dateTime = jsonReader.nextString();
+            if ("null".equals(dateTime)) return null;
+            return LocalDateTime.parse(dateTime, DATE_TIME_FORMATTER);
+        }
+    }
+}
