--- src/ru/yandex/practicum/exceptionPackage/ManagerSaveException.java
@@ -1,8 +1,0 @@
-package ru.yandex.practicum.exceptionPackage;
-
-public class ManagerSaveException extends RuntimeException {
-    public ManagerSaveException(final String message) {
-        super(message);
-    }
-}
-
