--- src/ru/yandex/practicum/exceptionPackage/TaskInputDateException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.exceptionPackage;
-
-public class TaskInputDateException extends Exception {
-    public TaskInputDateException(final String message) {
-        super(message);
-    }
-}
