--- src/ru/yandex/practicum/tasks/TaskStatuses.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.tasks;
-
-public enum TaskStatuses {
-    DONE,
-    IN_PROGRESS,
-    NEW
-}
