--- src/ru/yandex/practicum/tasks/TypeOfTask.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.tasks;
-
-public enum TypeOfTask {
-    TASK,
-    EPIC,
-    SUBTASK
-}
