--- src/ru/yandex/practicum/test/PrioritizedEndpointsTest.java
@@ -1,103 +1,0 @@
-package ru.yandex.practicum.test;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import ru.yandex.practicum.http.HttpTaskServer;
-import ru.yandex.practicum.tasks.Epic;
-import ru.yandex.practicum.tasks.Subtask;
-import ru.yandex.practicum.utils.Utils;
-
-import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.time.LocalDateTime;
-
-import static java.time.Month.AUGUST;
-
-public class PrioritizedEndpointsTest {
-    private HttpTaskServer httpTaskServer;
-    private HttpClient client;
-    private static final String DEFAULT_PRIORITIZED_URI = "http://localhost:8080/prioritized";
-    private static final Gson gson = new GsonBuilder()
-            .registerTypeAdapter(LocalDateTime.class, new Utils.LocalDateTimeAdapter())
-            .create();
-
-    @BeforeEach
-    public void startServer() throws IOException {
-        httpTaskServer = new HttpTaskServer();
-        client = HttpClient.newHttpClient();
-
-        //
-    }
-
-    @AfterEach
-    public void endServer() {
-        httpTaskServer.closeConnection();
-
-    }
-
-    @Test
-    public void getEmptyPrioritizedListTest() throws IOException, InterruptedException {
-        URI uri = URI.create(DEFAULT_PRIORITIZED_URI);
-        HttpRequest request = HttpRequest.newBuilder()
-                .GET()
-                .uri(uri)
-                .build();
-        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response.statusCode(), "Not right response");
-    }
-
-    @Test
-    public void getPrioritizedListTest() {
-        try {
-            URI uri1 = URI.create("http://localhost:8080/epics");
-            Epic epic1 = new Epic("Съездить в Москву", "обязательно до лета");
-            HttpRequest request1 = HttpRequest.newBuilder()
-                    .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic1)))
-                    .header("Content-Type", "application/json")
-                    .uri(uri1)
-                    .build();
-            HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
-            Assertions.assertEquals(200, response1.statusCode(), "Not right response");
-
-            URI uri2 = URI.create("http://localhost:8080/subtasks");
-            Subtask subtask11 = new Subtask("купить билеты", "дешёвые билеты", 1,
-                    LocalDateTime.of(2023, AUGUST, 28, 13, 0), 60);
-            HttpRequest request2 = HttpRequest.newBuilder()
-                    .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask11)))
-                    .header("Content-Type", "application/json")
-                    .uri(uri2)
-                    .build();
-            HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
-            Assertions.assertEquals(200, response2.statusCode(), "Not right response");
-
-            URI uri3 = URI.create("http://localhost:8080/subtasks");
-            HttpRequest request3 = HttpRequest.newBuilder()
-                    .GET()
-                    .uri(uri3)
-                    .build();
-            HttpResponse<String> response3 = client.send(request3, HttpResponse.BodyHandlers.ofString());
-            Assertions.assertEquals(200, response3.statusCode(), "Not right response");
-
-            URI uri = URI.create(DEFAULT_PRIORITIZED_URI);
-            HttpRequest request = HttpRequest.newBuilder()
-                    .GET()
-                    .uri(uri)
-                    .build();
-            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-            Assertions.assertEquals(200, response.statusCode(), "Not right response");
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            Assertions.fail();
-        }
-    }
-}
-
-
